+++ spinta/formats/html/commands.py
@@ -557,7 +557,7 @@ def prepare_dtype_for_response(
     action: Action,
     select: dict = None,
 ):
-    super_ = commands.prepare_dtype_for_response[Context, Format, Ref, dict]
+    super_ = commands.prepare_dtype_for_response[Context, Format, Ref, type(value)]
     value = super_(context, fmt, dtype, value, data=data, action=action, select=select)
     if value is None:
         return Cell('', color=Color.null)